+++ tests/pandas/test_agentset.py
